Added function to check validity of PID

+++ src/client.c
@@ -6,11 +6,37 @@
 /*   By: wheino <wheino@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/15 15:06:41 by wheino            #+#    #+#             */
-/*   Updated: 2025/07/16 12:59:45 by wheino           ###   ########.fr       */
+/*   Updated: 2025/07/16 13:18:30 by wheino           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minitalk.h"
+
+pid_t	check_pid(const char *pid_string)
+{
+	int		i;
+	pid_t	server_pid;
+
+	i = 0;
+	if (!pid_string || pid_string[0] == '\0')
+		return (ERROR);
+	while (pid_string[i] != '\0')
+	{
+		if (ft_isdigit(pid_string[i]) == 0)
+		{
+			ft_printf("ERROR: Invalid PID format.\n");
+			return (ERROR);
+		}
+		i++;
+	}
+	server_pid = ft_atoi(pid_string);
+	if (kill(server_pid, 0) == -1)
+	{
+		ft_printf("ERROR: Incorrect PID. PID %d does not exist.\n", server_pid);
+		return (ERROR);
+	}
+	return (server_pid);
+}
 
 int	main(int argc, char *argv[])
 {
@@ -18,10 +44,13 @@ int	main(int argc, char *argv[])
 	
 	if (argc != 3)
 	{
-		ft_printf("ERROR: Execute with: ./client [PID] [MSG]\n");
+		ft_printf("ERROR: Execute with: ./client [PID] [MSG].\n");
 		return (EXIT_FAILURE);
 	}
-	pid = ft_atoi(argv[1]);
+	pid = check_pid(argv[1]);
+	if (pid == ERROR)
+		return (EXIT_FAILURE);
 	kill(pid, SIGUSR1);
 	return (EXIT_SUCCESS);
 }
+
